+++ packages/studio-components/src/components/dropdown/pc-dropdown.style.scss
@@ -25,7 +25,7 @@ pc-dropdown {
   cursor: pointer;
   position: relative;
 
-  .rv-dr-root {
+  .pc-dr-root {
     padding: 5px 9px;
     padding-right: 35px;
     flex-wrap: wrap;
