+++ packages/studio-components/src/components/list/dropdown-list-item.style.scss
@@ -1,5 +1,5 @@
 $hv-color: rgba(black, 0.04);
-pc-list {
+pc-dropdown-list-item {
   overflow-x: hidden;
   overflow-y: auto;
   max-height: 100%;
