+++ packages/studio-components/src/components/list/dropdown-list-item.tsx
@@ -1,8 +1,8 @@
 import { Component, h, Prop, Event, EventEmitter, Method, Listen, State } from '@stencil/core';
-import { getItemLabel } from '../../utils/item.helpers';
+import { getItemLabel } from '../../utils/dropdown-list-item.helpers';
 
-@Component({ tag: 'pc-list', styleUrl: 'pc-list.style.scss' })
-export class PcDropdownList {
+@Component({ tag: 'pc-dropdown-list-item', styleUrl: 'dropdown-list-item.style.scss' })
+export class PcDropdownListItem {
   @State() currentItem = 0;
   private selectedEl: HTMLElement;
   /**
